[FLINK-26965][state] Extract PeriodicMaterializationManager into state-backend-common
